BUG/MEDIUM: stconn: Report blocked send if sends are blocked by an error
When some data must be sent to the endpoint but an error was previously
reported, nothing is performed and we leave. But, in this case, the SC is not
notified the sends are blocked.

It is indeed an issue if the endpoint reports an error after consuming all
data from the SC. In the endpoint the outgoing data are trashed because of
the error, but on the SC, everything was sent, even if an error was also
reported.

Because of this bug, it is possible to have outgoing data blocked at the SC
level but without any write timeout armed. In some cases, this may lead to
blocking conditions where the stream is never closed.

So now, when outgoing data cannot be sent because an previous error was
triggered, a blocked send is reported. This way, it is possible to report a
write timeout.

This patch should fix the issue #2754. It must be backported as far as 2.8.

(cherry picked from commit fbc3de6e9e59679d2e9ece3984ce31b6a7dd418f)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit f55508ab670a1b10fd4d338c1646fdaeb5ad27ad)
[cf: Only the connection part is applicable to the 2.9]
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit 69495df630d54ab9ae1ab0c465bd30bccae37b51)
[cf: channel_is_empty() is used to test if there are outgoing data]
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>